Creating shared LHFlib to be used for python hooks with ctypes
