+++ CMakeLists.txt
@@ -9,8 +9,9 @@ set (CMAKE_CXX_STANDARD 14)
 
 SET(CMAKE_CXX_FLAGS  "${CMAKE_CXX_FLAGS} ${GCC_COVERAGE_COMPILE_FLAGS}")
 SET(CMAKE_EXE_LINKER_FLAGS  "${CMAKE_EXE_LINKER_FLAGS} ${GCC_COVERAGE_LINK_FLAGS}")
+SET(CMAKE_POSITION_INDEPENDENT_CODE ON)
 
-project("LHF")
+project("LHF" VERSION 1.0.0 DESCRIPTION "Lightweight Homology Framework")
 
 # Include sub-projects
 add_subdirectory("LHFmain")
